feat: Implement linking media with posts

+++ app/controllers/admin/media_controller.rb
@@ -4,6 +4,8 @@ class Admin::MediaController < Admin::BaseController
 
   def create
     medium = Medium.create!(public_id: SecureRandom.uuid_v4)
+    Current.user.posts.find(params[:post_id]).post_media.create!(medium: medium) if params[:post_id]
+
     render json: {
       id: medium.id,
       presigned_url: backblaze_client.presigned_upload_url(key: medium.key)
@@ -12,8 +14,17 @@ class Admin::MediaController < Admin::BaseController
 
   def destroy
     medium = Medium.find(params[:id])
-    backblaze_client.delete_object(key: medium.key)
-    medium.destroy!
+    if medium.destroy
+      backblaze_client.delete_object(key: medium.key)
+      head :no_content
+    else
+      render json: { errors: medium.errors.full_messages }, status: :unprocessable_content
+    end
+  end
+
+  def unlink
+    post = Current.user.posts.find(params[:post_id])
+    post.post_media.find_by!(medium_id: params[:id]).destroy
     head :no_content
   end
